refactor(whispering): inline local engine dispatch in transcribe.ts
services.transcriptions.{whispercpp,parakeet,moonshine}.transcribe(...)
were thin shims: each ran the shared preflight, an engine-specific
peculiarity (whisper size check, moonshine variant regex), and then
called the typed Tauri command. Three files, three error types, one
registry indirection, for what is now a single switch on the FE.

Inline dispatchLocalTranscription into one switch over selectedService.
Absorb WhisperCppError.CorruptedModelFile and
MoonshineError.InvalidModelDirectoryName as variants of
LocalTranscriptionError so the local-failure error union stays one tag
family instead of three. Move the surviving local-transcription.ts up
one directory to drop the lone-file local/ subdir.

Net: -456 lines, no behavior change. The remaining engine differences
(file vs directory, size check, variant parse) live where they are
used, not in a per-engine namespace.
